requirements.txt
asgiref==3.2.10
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-bootstrap-modal-forms==1.4.1
django-crispy-forms==1.9.2
django-dynamic-models==0.0.5
django-heroku==0.3.1
django-tables2==2.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.9
iexfinance==0.4.3
joblib==0.15.1
numpy==1.18.4
pandas==1.0.4
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2018.5
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.14.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.9
whitenoise==5.2.0
